+++ src/view/backlogView.ts
@@ -151,7 +151,6 @@ export class ProductBacklogView extends BasesView implements BacklogViewHost {
 			treeEl: this.treeEl,
 		});
 		this.resize = new ResizePolicy(this, this.viewEl, this.treeEl, this.toolbarEl, () => this.rowCtx());
-		this.dnd.setupRootDropZone();
 		this.cardDnd = new CardDragController(this, this.viewEl);
 		this.treeEl.addEventListener('keydown', (evt) => handleProjectionKeydown(this, evt));
 		this.registerDomEvent(document, 'dragend', () => this.dnd.clearDragState());